requirements.txt
asyncpg==0.24.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
cycler==0.10.0
definitions==0.2.0
gevent==21.8.0
greenlet==1.1.1
idna==3.2
importlib-metadata==4.8.1
joblib==1.0.1
kiwisolver==1.3.1
lxml==4.6.3
matplotlib==3.3.4
mergesort==0.0.1
nltk==3.6.2
np==1.0.2
numpy==1.19.5
Pillow==8.3.1
pkg_resources==0.0.0
psycopg2-binary==2.9.1
pyparsing==2.4.7
python-dateutil==2.8.2
PyYAML==5.4.1
regex==2021.8.28
requests==2.26.0
sets==0.3.2
six==1.16.0
tqdm==4.62.2
typing-extensions==3.10.0.1
urllib3==1.26.6
zipp==3.5.0
zope.event==4.5.0
zope.interface==5.4.0
